(CUDA) agrega resultados CUDA. Mark IV

+++ nbody_cuda/src/include/nbody.h
@@ -19,7 +19,7 @@ struct Body {
 };
 
 void simulateNBodyCPU(Body* bodies, int n, int steps, float dt = 0.01f);
-void simulateNBodyCUDA(Body* h_bodies,  int steps, float dt, const char* kernelFilename, size_t localSize, int n);
+void simulateNBodyCUDA(Body* h_bodies,  int steps, float dt, const char* kernelFilename, int localSize, int n);
 void generateRandomBodies(Body* bodies, int n);
 
 #endif
